feat: customizable ban/warn keywords
